[PPAPI] Removed obsolete TODO in var.h (public C++ API).
The TODO says to remove a constructor when an old bug is fixed. The bug
is long since fixed. However, since removing a constructor would
potentially break existing third-party code, we should not remove it.
So, just removing the TODO.

Note: The existence of this TODO is preventing any changes to var.h from
going through the CQ.

BUG=314120

Review URL: https://codereview.chromium.org/142403006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@249661 4ff67af0-8c30-449e-8e8b-ad334ec8d88c