Split the PNaCl IRT shim into 3 pieces, and include one piece into IRT.
(*) A "for irt" library, that will get linked into the IRT.
That library just has the auto-generated pnacl_shim.c file.
Also modify ppapi_proxy to include a "irt_shim_ppapi.c",
which makes use of the pnacl_shim.c file to define a private
hook, parallel to the hook provided by irt_ppapi.c.

(*) A "for the browser" library, that the browser-tester
and the PNaCl packager will include for PNaCl-in-the-browser.
This is still called "libpnacl_irt_shim.a" so that we don't
need to change the in-browser linker's commandline. However,
it will be placed in a different directory so that the
Chrome PNaCl packager can pick it up separately from
the AOT library. This for-browser library uses the new hook.

(*) A "AOT" library, that will go into the SDK for offline
pexe -> nexe translation. Still called "libpnacl_irt_shim.a"
as well, confusingly enough, so that we don't need to change
the pnacl-nativeld.py commandlines. Placed in a different
directory from the for-browser library. Does not use the
new hook, so that it only depends on stable interfaces.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3783
R=binji@chromium.org, dmichael@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/164373010

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@255807 4ff67af0-8c30-449e-8e8b-ad334ec8d88c