[PPAPI] It is now possible to pass a FileSystem to JS in PostMessage.
Upon receiving a resource, if it is a file system, the V8 Var converter
automatically converts it into a Blink DOMFileSystem and makes it
available in the V8 JavaScript context.

This also adds the possibility of converting other resource types to
JavaScript objects in the future.

BUG=345158

Review URL: https://codereview.chromium.org/173143005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@255891 4ff67af0-8c30-449e-8e8b-ad334ec8d88c