Pepper: Move GetSandboxISA() to PPB_NaCl_Private.
This is part of a series of changes to move the "trusted plugin" code inside
ppapi/native_client. I was moving the logic for JsonManifest and
PNaClManifest when I noticed that PNaClManifest depends on GetSandboxISA(), and
moving that seemed big enough to justify its own change.

I think that I preserved the architecture logic in arch_x86/sandbox_isa.cc, but
it would merit some careful review.

BUG=239656

Review URL: https://codereview.chromium.org/193313002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@257196 4ff67af0-8c30-449e-8e8b-ad334ec8d88c