Pepper: Make PPB ReportLoadError set lastError.
This is part of a series of changes to move more error handling logic to
components/nacl/renderer from ppapi/native_client. I would have moved over some
of the console logging work as part of this change as well, but it started to
look pretty large.

BUG=239656
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/197873030

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@258050 4ff67af0-8c30-449e-8e8b-ad334ec8d88c