PPAPI/NaCl: Define NexeLoadManager class.
This adds an object that tracks the lifetime of the "Plugin" class in the
trusted plugin. State is kept in this class as long as the plugin is alive, to
make accounting much simpler for PPB_NaCl_Private.

This is derived from https://codereview.chromium.org/180213006/ but adds a
bugfix that initializes NexeLoadManager earlier.

BUG=239656

Review URL: https://codereview.chromium.org/208913002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@259362 4ff67af0-8c30-449e-8e8b-ad334ec8d88c