Pepper: Change uma presubmit to never block CQ.
This changes the presubmit rule for adding pepper interface hashes to UMA to a
notification instead of a warning. This is because the presubmit is a bit too
broad, and was blocking some changes from being submitted through the CQ.

Using a notification message should be fine, since the impact of missing a
histogram is relatively low (we'll just see a number for a while), and the hash
can always be added later.

BUG=351389
R=xhwang@chromium.org

Review URL: https://codereview.chromium.org/221973003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@261411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c