Pepper: Fix comments for PPB and PPP InputEvent.
The comments provided in PPB_InputEvent and PPP_InputEvent incorrectly describe
how handled events are delivered to handlers in web pages.

When a plugin handles an event, the event is not "bubbled" to any default
handlers. However, any listeners in a web page run before the event reaches the
plugin, so the plugin can't prevent any listeners from getting the event.

An attempt to make the implementation match the specification broke existing
plugins, so the only reasonable approach is to make the comments match the
behavior.

BUG=324352
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/228653006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@263034 4ff67af0-8c30-449e-8e8b-ad334ec8d88c