Pepper: Move PnaclOptions outside trusted plugin.
I ran into this while moving the Manifest code out of the trusted plugin and
figured this was big enough to stand on its own.

We have to create a PP_PNaClOptions struct instead of PnaclOptions so that the
objects of that type can be passed to functions in PPB_NaCl_Private.

BUG=239656

Review URL: https://codereview.chromium.org/235983020

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@264365 4ff67af0-8c30-449e-8e8b-ad334ec8d88c