Trusted plugin: Clean up manifest error reporting.
This streamlines some of the error reporting in the Manifest support classes in
the trusted plugin. Without this change, it's a bit hard to tell when we report
errors and when we don't; this makes some logic more explicit for when errors
are reported.

This will hopefully make reviewing the change that moves the manifest support
out of the trusted plugin easier.

BUG=353592

Review URL: https://codereview.chromium.org/266723004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@267624 4ff67af0-8c30-449e-8e8b-ad334ec8d88c