Pepper: Simplify manifest download logic.
This cleans up some of the trusted plugin logic for requesting a NaCl manifest
and removes some interfaces in PPB_NaCl_Private, providing better
encapsulation.

This code is easier to follow after this change, as well, since more of the
logic is unified in ppb_nacl_private_impl.cc. This is part of a larger effort
to remove the trusted plugin.

BUG=239656

Review URL: https://codereview.chromium.org/288773002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@270887 4ff67af0-8c30-449e-8e8b-ad334ec8d88c