Pepper: Nexe downloading out of the trusted plugin.
This moves the logic for nexe downloading outside the trusted plugin. This is
part of a series of changes to remove FileDownloader and its associated logic.

I've tried to make this change small so it will be easy to review. A followup
change will clean up some of the accounting logic post-download to make this
path simpler and easier to follow.

BUG=370556
R=bbudge@chromium.org

Review URL: https://codereview.chromium.org/276423003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@271090 4ff67af0-8c30-449e-8e8b-ad334ec8d88c