PPAPI: Make generator support "no_interface_string"
This is so I can introduce a new way to do PPP interfaces that are
passed to a related PPB interface, instead of using PPP_GetInterface.

Note this doesn't add support for the PNaCl shim yet, although it should
at least make it compile.

BUG=367896

Review URL: https://codereview.chromium.org/281203002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@271425 4ff67af0-8c30-449e-8e8b-ad334ec8d88c