Pepper: Preparation for async LoadAndStartNexe.
This is a preperatory change for moving nexe loading and starting from SRPC to
Chrome IPC. It's much simpler from an IPC and threading perspective if the new
IPC traffic is asynchronous. This change updates Plugin::LoadAndStartNexe() to
prepare for having an asynchronous interface in ServiceRuntime.

BUG=333950

Review URL: https://codereview.chromium.org/316013005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@275097 4ff67af0-8c30-449e-8e8b-ad334ec8d88c