Add IPC::ChannelProxy::Create() and IPC::SyncChannel::Create()
This change replaces constructors with Create() methods of
ChannelProxy and SyncChannel. This open the possibility to introduce
polymorphism to these classes.

This is a revision of r274310 (https://codereview.chromium.org/301973003/)
in which I added bunch of Create*() method variants.
The chagne was reverted. This change no longer does it and just keeps
using Channel::Mode to specify the channel type.

TEST=none
BUG=377980
R=darin@chromium.org,jam@chromium.org

Review URL: https://codereview.chromium.org/310853003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@275140 4ff67af0-8c30-449e-8e8b-ad334ec8d88c