Pepper: Refactor some coordinator histograms.
We have a bunch of histograms that log data at the same time as
ReportTranslationFinished(), so we may as well move some of that logic out of
the trusted plugin. We use a different time range for translation for other
metrics, so I had to add a "HistogramTimeTranslate" method in addition to the
existing ones for "small", "medium", and "large" times.

BUG=239656
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/307403005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@277073 4ff67af0-8c30-449e-8e8b-ad334ec8d88c