Pepper: Simplify OpenResource() for Non-SFI.
The Non-SFI implementation of OpenResource is pretty complicated. This is
because the logic to support that operation was entirely in the trusted
plugin. Now, we can perform the necessary logic entirely in Chromium, so
it can be made much simpler.

CQ_EXTRA_TRYBOTS=tryserver.chromium:linux_rel_precise32
BUG=239656

Review URL: https://codereview.chromium.org/339213003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@278859 4ff67af0-8c30-449e-8e8b-ad334ec8d88c