Move PPB_TrueTypeFont_Dev host from renderer to browser.
On Linux, we can now use FontConfig in the browser process since it
is thread-safe, so we don't need to use SandboxIPC from the renderer.
In the browser, the needed functionality is split off from
SandboxIPCHandler into a static function,
MatchFontFaceWithFallback in a new pair of files, font_utils_linux.*.

This change also moves any potentially blocking font creation and
reading to the browser's blocking thread pool.

I reworked the PepperTrueTypeFont base class to make Create simpler.
- Added an Initialize method which returns the font descriptor.
- Removed the Describe method.

I reworked the resource to delay calls to the host until we receive the desc
and initialization is complete. Describe will now wait until Initialize completes,
and the host uses a SequencedTaskRunner to serialize tasks, so Initialize
completes before GetTableTags and GetTable calls complete.

The Describe method can be implemented without IPC since we have the desc
on the plugin side.

BUG=382729

Review URL: https://codereview.chromium.org/337203003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@280107 4ff67af0-8c30-449e-8e8b-ad334ec8d88c