Make IPC macros work with -Wunused-local-typedefs.
The typedef can be used by the IPC_MESSAGE_HANDLER macro, but that macro isn't
always used when IPC_BEGIN_MESSAGE_MAP is used. The alternatives are to either
pass ClassName::HanderFunction to IPC_MESSAGE_HANDLER (like what one has to do
to use IPC_MESSAGE_FORWARD), or to mark the typedef in IPC_BEGIN_MESSAGE_MAP
as ALLOW_UNUSED so that it isn't warned on.

BUG=321833
TBR=brettw

Review URL: https://codereview.chromium.org/422083002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@286115 4ff67af0-8c30-449e-8e8b-ad334ec8d88c