Extract PepperFileSystemTypeToFileSystemType() from ppapi.
This function is called only from within content to convert from a ppapi
enum type to webkit fileapi type (this one will be come a content enum
type in later change).

So anticipating that change from webkit/ to content/ we moved this
function from ppapi to content to make this future change a little bit
easier.

And it is also nice to do now because it removes the last include of
webkit/ from ppapi making it much more clearer in term of deps.

BUG=338338
TEST=blink_tests
R=dmichael@chromium.org,jam@chromium.org

Review URL: https://codereview.chromium.org/424213006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@286776 4ff67af0-8c30-449e-8e8b-ad334ec8d88c