Pepper: Remove trusted plugin loadable module.
This change links the trusted plugin into the renderer, making it easier for
the trusted plugin to use libraries in Chromium. This removes the
ppGoogleNaClPlugin loadable module file on various platforms.

This is part of a larger effort to remove the "trusted plugin" used to
bootstrap NaCl plugins.

It introduces an "internal_module" interface for setting the value returned by
pp::Module::Get(). This is so that both the trusted plugin and the remoting
plugin can be linked into the renderer. However, I believe this is safe because
the trusted plugin runs in-process and the remoting plugin runs out-of-process.

BUG=394497
R=dmichael@chromium.org, mseaborn@chromium.org, phajdan.jr@chromium.org, thestig@chromium.org, wez@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=287071

Review URL: https://codereview.chromium.org/397243004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@287580 4ff67af0-8c30-449e-8e8b-ad334ec8d88c