Pepper: Fix caching for translated nexes.
A refactor of FileDownloader (r284961) broke the caching behavior for nexes
that were translated from pexes; the cache would always end up with entries of
length 0, breaking future loads of the pexe. This is the translated nexe was
written to the wrong file handle, preventing caching of the result.

This change causes the translated nexe to be written to the correct file
handle, fixing the caching behavior.

I manually tested this change on the "game of life" demo at
https://gonativeclient.appspot.com/demo/life, verifying that a cached
translated nexe for the pexe was successfully loaded.

BUG=401121

Review URL: https://codereview.chromium.org/447013002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@287912 4ff67af0-8c30-449e-8e8b-ad334ec8d88c