ppapi: Add way to ignore presubmit warning about including in NaCl SDK.
The ppapi presubmit checks that a modified interface exists in the NaCl SDK.
But not all interfaces should be included. It is currently a warning, but that
will prevent the CQ from submitting the CL.

Now we can add that file to an "IGNORE_LIST" and it will no longer emit a
presubmit warning.

BUG=403713
R=bbudge@chromium.org
TEST=modified pp_video_dev.h, ran presubmits, got warning. Added to ignore
list, ran again. No warning.

Review URL: https://codereview.chromium.org/475763003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@289734 4ff67af0-8c30-449e-8e8b-ad334ec8d88c