Pepper: Make pending_sync_msgs_ local to IO thread.
Adding support for PpapiHostMessage_OpenResource to SFI mode is tricky.
Validating the file tokens requires that a message to be sent to the browser
(!) to validate that information in the case of a compromised renderer.

This change is one step to allowing processing for the reply of the
PpapiHostMessage_OpenResource message to send a message to the browser for
constructing a NaClDesc properly.

This makes the encapsulation of NaClMessageScanner a bit worse, but also
reduces the amount of code for which the critical section under lock_ must be
held.

BUG=394130

Review URL: https://codereview.chromium.org/472073003

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@291234 4ff67af0-8c30-449e-8e8b-ad334ec8d88c