PPAPI: make generator output a pointer to an array correctly.
Originally an IDL argument like "[out] int[] foo" will generate
int* foo[]. It means an array of pointers to integers and what we
want is a pointer to an array of integers. Change the output to
int** foo to be not confusing.

BUG=387547
TEST=Run generator and build chrome.

Review URL: https://codereview.chromium.org/490433002

Cr-Original-Commit-Position: refs/heads/master@{#291349}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab8bcef370b99db6414eee827ba845a44d683380