Fix use-after-free in StartSelLdrOnMainThread.
There is a use-after free in StartSelLdrOnMainThread(), where that
function may be called on the main thread after service_runtime is
destroyed. This change binds service_runtime to subprocess later,
preventing the deletion that was happening before a variable use.

Here's the use-after-free path: Start at Plugin::LoadHelperNaClModule,
where a nacl::scoped_ptr<NaClSubprocess> is allocated on the stack.

Plugin::LoadHelperNaClModuleInternal is called next, which allocates a
ServiceRuntime and sets it as the ServiceRuntime for the above
NaClSubprocess, hands a pointer to that ServiceRuntime to a Callback to
StartSelLdrOnMainThread, and calls WaitForSelLdrStart().

If WaitForSelLdrStart() times out, LoadHelperNaClModuleInternal returns
NULL, as does LoadHelperNaClModule, and the NaClSubprocess is destroyed.
The destructor for NaClSubprocess deletes its service_runtime_ member
(while our in-flight callback is still pointing to it).

Then, suppose that StartSelLdrOnMainThread finally gets a chance to run
on the main thread. The ServiceRuntime is already deleted, and this is
revealed when the process crashes when trying to delete what's pointed
to by subprocess_ inside that ServiceRuntime.

BUG=403942

Review URL: https://codereview.chromium.org/844793002

Cr-Original-Commit-Position: refs/heads/master@{#310550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c8d343e6fca46a7ed092926439b7b5b1b317318