NaCl: Remove disabled test, fault_injection/fault_pm_nameservice_test.cc
This has been disabled since September 2011, when it got disabled for
being flaky by this change: http://codereview.chromium.org/7833022

fault_pm_nameservice_test.cc depended on the SRPC services
"ManifestNameService" (which we are removing) and "SecureRandom"
(which we have already removed).

The use of NACL_FAULT_INJECTION is pretty obscure, and is not
something that we're going to try to re-enable.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3864
TEST=none

Review URL: https://codereview.chromium.org/842363002

Cr-Original-Commit-Position: refs/heads/master@{#310970}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6daf793a020c57f7b505d76df87a7a0a5767ef2