Remove IPC logging from Nacl IRT's dispatcher
In the GN debug build, the IPC logging classes conflict with the ones in
content/common, leading to duplicate symbols. In the component build, these end
up in different shared libraries so there is no conflict, and in release the
logging is disabled.

I believe this doesn't come up in GYP due to the static libraries in there
which I assume meant that the content object files were never pulled in to the
target. For some reason this regressed in the GN build recently, but I could
not find related changes.

Review URL: https://codereview.chromium.org/1159553008

Cr-Original-Commit-Position: refs/heads/master@{#332156}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 574ee19dc3a25a5b6b127e7991a436a831a30a10